Data Loader script
Pre-Loads the data upon app startup. The data is stored in a javascript object that we can access from web3. You can run it in the terminal with "node data-loader.js". The Errors that print are nox values in Row 758 that arent accepted by the Decimal instance. They are skipped and I decrement the tablesize by 1, as a temporary solution. This errror only occurs in datatables 2003-2006.