Provable Implementation
We have 10 request urls. The provable_query did not recognize the url value if I tried to submit a string variable into instead of the hardcoded values. So i just created 10 request functions that request the mean from their respective paths. I deployed this contract and the address of it is at the top of the contract. You can test it out on Remix.

TODOs include:
1) convert the response from a string into uint (provables response will always be a string so, there is no way to do this offchain). I will prob have to change the data in the API so it returns a rounded value, since Solidity doesnt support floating points. We could also use a library to replicate floating point numbers, although it might be expensive.

2) Route index to the proper updateMean method in Oracle.sol. So that we can simulate the value changing overtime. Provable allows for call delays and recursive calls, so we can have it call the methods automatically after a set amount of time. I think the max time limit is 60 days, but it should be enough to prove concept in the prototype and demo.