Node Backend
Example Contract Address:
0xEEE5b03a866426Da5dc6d9142eD06ce859bFe3B1

The backend starts a cron job to call the updateTotalOwed function every 5 minutes. Change it to 1 minute to test. There is a Receipt error when making too many transactions in a row on ropsten. The transaction will still be queued though, and will execute by the transaction nonce. Tested on the Example Contract (Example.sol). NOx Values are sent onchain.