Powergate v1 -Beta
There are lots of things i need to fix, but I ignored some of them just to get the demo to work. One of the problems is that the powergate instance does not keep the array of CIDs in the same order as the order that they were uploaded in.